Updated ISA,added bytecode concreate format,refactored assembler

+++ assembler/codegen.c
@@ -1,7 +1,9 @@
 #include "dynamic_array.h"
-#include "parser.h"
-#include "stdio.h"
+#include <stdio.h>
 #include "../format.h"
+#include "../bytecode/bytecode_format.h"
+#include "../isa/isa_table.h"
+
 
 
 void codegen(char *output_file, DynamicArray *instructions)
@@ -10,7 +12,7 @@ void codegen(char *output_file, DynamicArray *instructions)
   fwrite(MAGIC, 1, MAGIC_SIZE, output);
   fwrite(&instructions->length, 1, sizeof(uint16_t), output);
 
-  for(int i = 0; i < instructions->length; i++)
+  for(size_t i = 0; i < instructions->length; i++)
   {
     Instruction *instr = (Instruction *) get_element(instructions, i);
     Instr_def const *instr_def = instr_def_by_type(instr->type);
